requirements-dev.txt
# Development Dependencies for Gestionale Comune
# Install with: pip install -r requirements-dev.txt

# Code Quality & Linting
flake8==6.1.0                   # Python linting
black==23.9.1                   # Code formatting
isort==5.12.0                   # Import sorting
pylint==3.0.2                   # Advanced Python linting

# Testing
pytest==7.4.3                   # Modern testing framework
pytest-django==4.7.0            # Django integration for pytest
pytest-cov==4.1.0               # Coverage reporting
factory-boy==3.3.0              # Test data factories
model-bakery==1.17.0            # Django model factories

# Development Tools
django-extensions==3.2.3        # Useful Django commands
django-debug-toolbar==4.2.0     # Debug toolbar for development
ipython==8.17.2                 # Enhanced Python shell
pre-commit==3.5.0               # Git hooks for code quality

# Documentation
sphinx==7.2.6                   # Documentation generator
sphinx-rtd-theme==1.3.0         # ReadTheDocs theme

# Database Tools
django-migration-linter==5.0.0  # Migration linting

# Security
bandit==1.7.5                   # Security linting
safety==2.3.5                   # Dependency vulnerability scanner